made frasco-models transaction based for all backends
even if a backend does not use transaction and just save/deletes directly,
all model objects manipulations will be wrapped in a transaction.
this allows better support for backends which need transactions.

whether to auto commit calls to save()/add() outside of a transaction or
commit on teardown is left to the backend.

features developer should wrap any model manipulations in a transaction
as it may get lost if the backend does not commit on teardown.